+++ changelog.md
@@ -1,2 +1,2 @@
 ### Formations 1.0.2
-- Initial release of Formations for Minecraft 1.20.5 & 1.20.6
+- Initial release of Formations for Minecraft 1.21